style: #169 tag 컴포넌트 gray 수정
